std: add container::{Seq, MutableSeq}
