+++ src/libstd/prelude.rs
@@ -51,7 +51,8 @@ pub use c_str::ToCStr;
 pub use char::Char;
 pub use clone::{Clone, DeepClone};
 pub use cmp::{Eq, ApproxEq, Ord, TotalEq, TotalOrd, Ordering, Less, Equal, Greater, Equiv};
-pub use container::{Container, NewContainer, Mutable, Map, MutableMap, Set, MutableSet};
+pub use container::{Container, NewContainer, Mutable, Seq, MutableSeq, Map, MutableMap};
+pub use container::{Set, MutableSet};
 pub use default::Default;
 pub use from_str::FromStr;
 pub use hash::Hash;
